requirements-lite.txt
# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0

# Input Validation
marshmallow==3.20.1
email-validator==2.1.0

# HTTP Client & File Handling
Werkzeug==3.0.1

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0

# Date & Time
python-dateutil==2.8.2

# Production WSGI Server
waitress==2.1.2 